add new lockstack delcaration to globals and utility binaries
