+++ src/threaddeadlock.h
@@ -8,7 +8,15 @@
 
 #include <boost/lexical_cast.hpp>
 #include <boost/thread.hpp>
+#include <inttypes.h>
+#include <map>
+#include <memory>
+#include <mutex>
+#include <set>
 #include <string>
+#include <string>
+#include <utility>
+#include <vector>
 
 #include "utilstrencodings.h"
 
@@ -72,6 +80,43 @@ private:
     bool fExclusive; // signifies Exclusive Ownership, this is always true for a CCriticalSection
     bool fWaiting; // determines if lock is held or is waiting to be held
 };
+
+// pair ( cs : lock location )
+typedef std::pair<void *, CLockLocation> LockStackEntry;
+typedef std::vector<LockStackEntry> LockStack;
+
+// cs : set of thread ids
+typedef std::map<void *, std::set<uint64_t> > ReadLocksHeld;
+// cs : set of thread ids
+typedef std::map<void *, std::set<uint64_t> > WriteLocksHeld;
+
+// cs : set of thread ids
+typedef std::map<void *, std::set<uint64_t> > ReadLocksWaiting;
+// cs : set of thread ids
+typedef std::map<void *, std::set<uint64_t> > WriteLocksWaiting;
+
+// thread id : vector of locks held (both shared and exclusive, waiting and held)
+typedef std::map<uint64_t, LockStack> LocksHeldByThread;
+
+
+struct LockData
+{
+    // Very ugly hack: as the global constructs and destructors run single
+    // threaded, we use this boolean to know whether LockData still exists,
+    // as DeleteLock can get called by global CCriticalSection destructors
+    // after LockData disappears.
+    bool available;
+    LockData() : available(true) {}
+    ~LockData() { available = false; }
+    ReadLocksWaiting readlockswaiting;
+    WriteLocksWaiting writelockswaiting;
+
+    ReadLocksHeld readlocksheld;
+    WriteLocksHeld writelocksheld;
+    LocksHeldByThread locksheldbythread;
+    std::mutex dd_mutex;
+};
+extern LockData lockdata;
 
 void push_lock(void *c, const CLockLocation &locklocation, LockType type, bool isExclusive, bool fTry);
 void DeleteCritical(void *cs);
